+++ lib/rspec/core/runner.rb
@@ -8,7 +8,12 @@ module RSpec
       def self.autorun
         return if autorun_disabled? || installed_at_exit? || running_in_drb?
         @installed_at_exit = true
-        at_exit { exit(run(ARGV, $stderr, $stdout).to_i) }
+        # The order here is important, the at_exit handler must be installed before anyone else gets a chance
+        # to install their own, that way we can be assured that our exit will be last to run (at_exit stacks).
+        # see https://github.com/jnicklas/capybara/issues/178
+        exit_code = nil
+        at_exit { exit exit_code if exit_code && exit_code != 0 }
+        exit_code = run(ARGV, $stderr, $stdout).to_i
       end
       AT_EXIT_HOOK_BACKTRACE_LINE = "#{__FILE__}:#{__LINE__ - 2}:in `autorun'"